+++ example/lib/affise/affise_widget.dart
@@ -52,8 +52,6 @@ class _AffiseWidgetState extends State<AffiseWidget> {
                   // Send event
                   items[index].send();
                   // or
-                  // Affise.sendEvent(items[index]);
-                  // or
                   // items[index].sendNow(() {
                   //   debugPrint("success: ${items[index].getName()}");
                   // }, (status) {