+++ lib/module/attribution/affise_attribution_module_api.dart
@@ -10,12 +10,4 @@ abstract class AffiseAttributionModulesApi {
   void getStatus(AffiseModules module, OnKeyValueCallback callback);
 
   Future<List<AffiseModules>> getModulesInstalled();
-
-  void linkResolve(String url, AffiseLinkCallback callback);
-
-  void fetchProducts(List<String> ids, AffiseResultCallback<AffiseProductsResult> callback);
-
-  void purchase(AffiseProduct product, AffiseProductType type, AffiseResultCallback<AffisePurchasedInfo> callback);
 }
-
-