+++ lib/module/attribution/affise_ios_api.dart
@@ -6,8 +6,4 @@ abstract class AffiseIOSApi {
   void registerAppForAdNetworkAttribution(ErrorCallback completionHandler);
 
   void updatePostbackConversionValue(int fineValue, CoarseValue coarseValue, ErrorCallback completionHandler);
-
-  void getReferrerOnServer(ReferrerCallback callback);
-
-  void getReferrerOnServerValue(ReferrerKey key, ReferrerCallback callback);
 }